Fix for version/author field mixup
Currently (under Sat 6.0.4.) the hammer command to list the modules

hammer --csv puppet-module list --repository-id ${PUPPET_REPO_ID}"...

computes something like

ssh@2.4.0

this is then used as input for --name and --author, which leads to
obvious problems.

The change computes something like

ssh@ACME_corp

I guess this is probably due to some changes in the output of the
hammer command and thus I'm not sure if this breaks previous versions.

but later on this is passed to --