reboot only works for running hosts, we have to start the shutoff ones
When a host that should be rebuilt is shut off, "hammer host reboot" will
fail with an missleading message:
 Call to virDomainReboot failed:
  Requested operation is not valid:
   Reboot is not supported without the JSON monitor

While I think this is a bug in the libvirt API -- a reboot of a shut off
domain should just boot it, as it happens with services all the time --
we can and should just fix it in our scripts.

This commit checks the status of the host and acts accordingly, depending
on the result (running → reboot, shutoff → start).